refactor subnav
